Use getMethodNameFromConstantPool to obtain method name
Previously, getMethodNameAndSignatureFromConstantPool was used
which only returned the signature. In the function
shouldCompileTimeResolveMethod, we check if the method is
invokeBasic, in which case it must always be compile-time resolved.
However, invokeBasic could never be identified as we were not
getting the method name or the correct method name length from
getMethodNameAndSignatureFromConstantPool.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>